split table component
